% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dyadic_univariate_example}
\alias{dyadic_univariate_example}
\title{Synthetic univariate dyadic sequence example}
\format{
A data frame with 90 rows and 3 columns:
\describe{
\item{time}{Index of the measurement occasion.}
\item{FM}{Integer state for the first member, taking values 1 or 2.}
\item{SM}{Integer state for the second member, taking values 1 or 2.}
}
}
\usage{
dyadic_univariate_example
}
\description{
A synthetic dyadic sequence with 90 observations, designed for package
workflow examples.
}
\details{
The package workflow classifies this example as \code{PM (A3)}
using \code{\link{univariatePattern}} with \code{states = 2}.
}
\keyword{datasets}
